+++ R/qts-hclust.R
@@ -38,21 +38,30 @@ hclust.default <- function(x,
 #' @export
 #' @rdname hclust
 hclust.qts_sample <-function(x,
-                             metric = c("l2", "pearson"),
+                             metric = c("l2", "normalized_l2", "pearson"),
                              linkage_criterion = c("complete", "average", "single", "ward.D2"),
                              n_clusters = 1L,
-                             warping_class = c("affine", "dilation", "none", "shift", "srsf"),
+                             is_domain_interval = FALSE,
+                             transformation = c("identity", "srvf"),
+                             warping_class = c("none", "shift", "dilation", "affine", "bpd"),
                              centroid_type = "mean",
                              cluster_on_phase = FALSE,
                              ...) {
   call <- rlang::call_match(defaults = TRUE)
+
   call_args <- rlang::call_args(call)
-  metric <- rlang::arg_match(metric)
-  linkage_criterion <- rlang::arg_match(linkage_criterion)
+
+  transformation <- rlang::arg_match(transformation)
+  call_args$transformation <- transformation
+
   warping_class <- rlang::arg_match(warping_class)
-  call_args$metric <- metric
-  call_args$linkage_criterion <- linkage_criterion
   call_args$warping_class <- warping_class
+
+  metric <- rlang::arg_match(metric)
+  call_args$metric <- metric
+
+  linkage_criterion <- rlang::arg_match(linkage_criterion)
+  call_args$linkage_criterion <- linkage_criterion
 
   l <- prep_data(x)
 
@@ -60,6 +69,8 @@ hclust.qts_sample <-function(x,
     x = l$grid,
     y = l$values,
     n_clusters = n_clusters,
+    is_domain_interval = is_domain_interval,
+    transformation = transformation,
     warping_class = warping_class,
     centroid_type = centroid_type,
     metric = metric,
@@ -69,7 +80,7 @@ hclust.qts_sample <-function(x,
   )
 
   res <- list(
-    qts_aligned = purrr::map(1:l$N, \(.id) {
+    qts_aligned = lapply(1:l$N, \(.id) {
       exp(as_qts(tibble::tibble(
         time = out$aligned_grids[.id, ],
         w    = 0,
@@ -78,7 +89,7 @@ hclust.qts_sample <-function(x,
         z    = out$original_curves[.id, 3, ]
       )))
     }),
-    qts_centers = purrr::map(1:out$n_clusters, \(.id) {
+    qts_centers = lapply(1:out$n_clusters, \(.id) {
       exp(as_qts(tibble::tibble(
         time = out$center_grids[.id, ],
         w    = 0,