+++ R/qts-transformations.R
@@ -115,3 +115,31 @@ qts2aats <- function(x) {
   names(out) <- c("time", "angle", "ux", "uy", "uz")
   out
 }
+
+#' QTS Transformation to Roll-Pitch-Yaw Time Series
+#'
+#' This function converts a quaternion time series into its roll-pitch-yaw
+#' angles representation.
+#'
+#' @param x An object of class [qts].
+#'
+#' @return A time series stored as a [tibble::tibble] with columns `time`,
+#'   `roll`, `pitch` and `yaw` containing the roll, pitch and yaw angles
+#'   representation of the input quaternions.
+#'
+#' @export
+#' @examples
+#' qts2rpyts(vespa64$igp[[1]])
+qts2rpyts <- function(x) {
+  if (!is_qts(x))
+    cli::cli_abort("The input argument {.arg x} should be of class {.cls qts}.")
+  out <- qts2rpyts_impl(x)
+  names(out) <- c("time", "roll", "pitch", "yaw")
+  out
+}
+
+rpyts2qts <- function(x) {
+  x |>
+    rpyts2qts_impl() |>
+    as_qts()
+}