+++ R/utils.R
@@ -3,8 +3,8 @@ prep_data <- function(x) {
     cli::cli_abort("The input argument {.arg x} should be of class {.cls qts_sample}.")
 
   q_list <- log(x)
-  q_list <- purrr::map(q_list, \(.x) rbind(.x$x, .x$y, .x$z))
-  t_list <- purrr::map(x, "time")
+  q_list <- lapply(q_list, \(.x) rbind(.x$x, .x$y, .x$z))
+  t_list <- lapply(x, \(.x) .x$time)
 
   # Prep data
   N <- length(q_list)
@@ -25,4 +25,20 @@ prep_data <- function(x) {
   }
 
   list(grid = grid, values = values, N = N, L = L, P = P)
+}
+
+transpose <- function(x) {
+  n <- length(x[[1]])
+  lens <- lengths(x)
+  if (any(lens != n))
+    cli::cli_abort("All elements of the input list should have the same length.")
+  lapply(1:n, \(i) sapply(x, \(.x) .x[i]))
+}
+
+transpose_flatten <- function(x) {
+  n <- length(x[[1]])
+  lens <- lengths(x)
+  if (any(lens != n))
+    cli::cli_abort("All elements of the input list should have the same length.")
+  lapply(1:n, \(i) sapply(x, \(.x) .x[[i]]))
 }